fix(max_tile_size): Changed the max_tile_size typing to float.

+++ src/geefetch/cli/omegaconfig.py
@@ -33,14 +33,14 @@ class GEEConfig:
         Your GEE id, to connect to the API.
 
         .. see also:: https://developers.google.com/earth-engine/apidocs/ee-initialize
-    max_tile_size : int
+    max_tile_size : float
         Size constraint in MB for the request sent to GEE. This is heuristical and depends
         in general on what satellite you are interested in.
         Decrease if User Memory Excess Error, but choose highest possible otherwise. Defaults is 10.
     """
 
     ee_project_id: str = "my-ee-project"
-    max_tile_size: int = 10
+    max_tile_size : float = 10
 
 
 @dataclass
