+++ src/geefetch/data/get.py
@@ -170,7 +170,7 @@ def download_time_series(
     crs: CRS | None = None,
     resolution: int = 10,
     tile_shape: int = 500,
-    max_tile_size: int = 5,
+    max_tile_size : float = 5,
     satellite_get_kwargs: dict[str, Any] | None = None,
     satellite_download_kwargs: dict[str, Any] | None = None,
     check_clean: bool = True,
@@ -204,7 +204,7 @@ def download_time_series(
         Resolution of the downloaded data, in meters. Defaults to 10.
     tile_shape : int
         Side length of a downloaded chip, in pixels. Defaults to 500.
-    max_tile_size : int
+    max_tile_size : float
         Parameter adjusting the memory consumption in Google Earth Engine, in Mb.
         Choose the highest possible that doesn't raise a User Memory Excess error.
         Defaults to 10 Mb.
@@ -293,7 +293,7 @@ def download(
     crs: CRS | None = None,
     resolution: int = 10,
     tile_shape: int = 500,
-    max_tile_size: int = 5,
+    max_tile_size : float = 5,
     satellite_get_kwargs: dict[str, Any] | None = None,
     satellite_download_kwargs: dict[str, Any] | None = None,
     check_clean: bool = True,
@@ -328,7 +328,7 @@ def download(
         Resolution of the downloaded data, in meters. Defaults to 10.
     tile_shape : int
         Side length of a downloaded chip, in pixels. Defaults to 500.
-    max_tile_size : int
+    max_tile_size : float
         Parameter adjusting the memory consumption in Google Earth Engine, in Mb.
         Choose the highest possible that doesn't raise a User Memory Excess error.
         Defaults to 10 Mb.
@@ -473,7 +473,7 @@ def download_gedi(
     crs: CRS | None = None,
     resolution: int = 10,
     tile_shape: int = 500,
-    max_tile_size: int = 5,
+    max_tile_size : float = 5,
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
@@ -501,7 +501,7 @@ def download_gedi(
         Resolution of the downloaded data, in meters. Defaults to 10.
     tile_shape : int
         Side length of a downloaded chip, in pixels. Defaults to 500.
-    max_tile_size : int
+    max_tile_size : float
         Parameter adjusting the memory consumption in Google Earth Engine, in Mb.
         Choose the highest possible that doesn't raise a User Memory Excess error. Defaults to 10.
     composite_method : CompositeMethod
@@ -615,7 +615,7 @@ def download_s1(
     crs: CRS | None = None,
     resolution: int = 10,
     tile_shape: int = 500,
-    max_tile_size: int = 5,
+    max_tile_size : float = 5,
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
@@ -645,7 +645,7 @@ def download_s1(
         Resolution of the downloaded data, in meters. Defaults to 10.
     tile_shape : int
         Side length of a downloaded chip, in pixels. Defaults to 500.
-    max_tile_size : int
+    max_tile_size : float
         Parameter adjusting the memory consumption in Google Earth Engine, in Mb.
         Choose the highest possible that doesn't raise a User Memory Excess error. Defaults to 10.
     composite_method : CompositeMethod
@@ -718,7 +718,7 @@ def download_s2(
     crs: CRS | None = None,
     resolution: int = 10,
     tile_shape: int = 500,
-    max_tile_size: int = 5,
+    max_tile_size : float = 5,
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
@@ -749,7 +749,7 @@ def download_s2(
         Resolution of the downloaded data, in meters. Defaults to 10.
     tile_shape : int
         Side length of a downloaded chip, in pixels. Defaults to 500.
-    max_tile_size : int
+    max_tile_size : float
         Parameter adjusting the memory consumption in Google Earth Engine, in Mb.
         Choose the highest possible that doesn't raise a User Memory Excess error. Defaults to 10.
     composite_method : CompositeMethod
@@ -813,7 +813,7 @@ def download_dynworld(
     crs: CRS | None = None,
     resolution: int = 10,
     tile_shape: int = 500,
-    max_tile_size: int = 5,
+    max_tile_size : float = 5,
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
@@ -842,7 +842,7 @@ def download_dynworld(
         Resolution of the downloaded data, in meters. Defaults to 10.
     tile_shape : int
         Side length of a downloaded chip, in pixels. Defaults to 500.
-    max_tile_size : int
+    max_tile_size : float
         Parameter adjusting the memory consumption in Google Earth Engine, in Mb.
         Choose the highest possible that doesn't raise a User Memory Excess error. Defaults to 10.
     composite_method : CompositeMethod
@@ -899,7 +899,7 @@ def download_landsat8(
     crs: CRS | None = None,
     resolution: int = 30,
     tile_shape: int = 500,
-    max_tile_size: int = 5,
+    max_tile_size : float = 5,
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
@@ -928,7 +928,7 @@ def download_landsat8(
         Resolution of the downloaded data, in meters. Defaults to 30.
     tile_shape : int
         Side length of a downloaded chip, in pixels. Defaults to 500.
-    max_tile_size : int
+    max_tile_size : float
         Parameter adjusting the memory consumption in Google Earth Engine, in Mb.
         Choose the highest possible that doesn't raise a User Memory Excess error. Defaults to 10.
     composite_method : CompositeMethod
@@ -983,7 +983,7 @@ def download_palsar2(
     crs: CRS | None = None,
     resolution: int = 30,
     tile_shape: int = 500,
-    max_tile_size: int = 5,
+    max_tile_size : float = 5,
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
@@ -1014,7 +1014,7 @@ def download_palsar2(
         Resolution of the downloaded data, in meters. Defaults to 30.
     tile_shape : int
         Side length of a downloaded chip, in pixels. Defaults to 500.
-    max_tile_size : int
+    max_tile_size : float
         Parameter adjusting the memory consumption in Google Earth Engine, in Mb.
         Choose the highest possible that doesn't raise a User Memory Excess error. Defaults to 10.
     composite_method : CompositeMethod
@@ -1074,7 +1074,7 @@ def download_nasadem(
     crs: CRS | None = None,
     resolution: int = 10,
     tile_shape: int = 500,
-    max_tile_size: int = 5,
+    max_tile_size : float = 5,
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Polygon | None = None,
@@ -1099,7 +1099,7 @@ def download_nasadem(
         Resolution of the downloaded data, in meters. Defaults to 10.
     tile_shape : int
         Side length of a downloaded chip, in pixels. Defaults to 500.
-    max_tile_size : int
+    max_tile_size : float
         Parameter adjusting the memory consumption in Google Earth Engine, in Mb.
         Choose the highest possible that doesn't raise a User Memory Excess error. Defaults to 10.
     composite_method : CompositeMethod
@@ -1155,7 +1155,7 @@ def download_custom(
     crs: CRS | None = None,
     resolution: int = 10,
     tile_shape: int = 500,
-    max_tile_size: int = 10,
+    max_tile_size : float = 10,
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Polygon | None = None,
@@ -1185,7 +1185,7 @@ def download_custom(
         Resolution of the downloaded data, in meters. Defaults to 10.
     tile_shape : int
         Side length of a downloaded chip, in pixels. Defaults to 500.
-    max_tile_size : int
+    max_tile_size : float
         Parameter adjusting the memory consumption in Google Earth Engine, in Mb.
         Choose the highest possible that doesn't raise a User Memory Excess error. Defaults to 10.
     composite_method : CompositeMethod
